curves.py mvp
